+++ packages/client-core/src/systems/WidgetUISystem.ts
@@ -45,10 +45,7 @@ import { V_001, V_010 } from '@etherealengine/spatial/src/common/constants/MathC
 import { InputSourceComponent } from '@etherealengine/spatial/src/input/components/InputSourceComponent'
 import { XRStandardGamepadButton } from '@etherealengine/spatial/src/input/state/ButtonState'
 import { VisibleComponent, setVisibleComponent } from '@etherealengine/spatial/src/renderer/components/VisibleComponent'
-import {
-  ComputedTransformComponent,
-  setComputedTransformComponent
-} from '@etherealengine/spatial/src/transform/components/ComputedTransformComponent'
+import { ComputedTransformComponent } from '@etherealengine/spatial/src/transform/components/ComputedTransformComponent'
 import { TransformComponent } from '@etherealengine/spatial/src/transform/components/TransformComponent'
 import { ReferenceSpace, XRState, isMobileXRHeadset } from '@etherealengine/spatial/src/xr/XRState'
 import {
@@ -188,10 +185,13 @@ const execute = () => {
   } else {
     if (!hasComponent(widgetMenuUI.entity, ComputedTransformComponent)) {
       setComponent(widgetMenuUI.entity, EntityTreeComponent, { parentEntity: Engine.instance.originEntity })
-      setComputedTransformComponent(widgetMenuUI.entity, Engine.instance.viewerEntity, () => {
-        const camera = getComponent(Engine.instance.viewerEntity, CameraComponent)
-        const distance = camera.near * 1.1 // 10% in front of camera
-        ObjectFitFunctions.attachObjectInFrontOfCamera(widgetMenuUI.entity, 0.2, distance)
+      setComponent(widgetMenuUI.entity, ComputedTransformComponent, {
+        referenceEntities: [Engine.instance.viewerEntity],
+        computeFunction: () => {
+          const camera = getComponent(Engine.instance.viewerEntity, CameraComponent)
+          const distance = camera.near * 1.1 // 10% in front of camera
+          ObjectFitFunctions.attachObjectInFrontOfCamera(widgetMenuUI.entity, 0.2, distance)
+        }
       })
     }
   }
