+++ packages/engine/src/interaction/components/InteractableComponent.ts
@@ -25,6 +25,7 @@ Ethereal Engine. All Rights Reserved.
 
 import { isClient } from '@etherealengine/common/src/utils/getEnvironment'
 import {
+  ECSState,
   Engine,
   Entity,
   EntityUUID,
@@ -38,23 +39,31 @@ import {
   useEntityContext,
   useOptionalComponent
 } from '@etherealengine/ecs'
-import { defineComponent } from '@etherealengine/ecs/src/ComponentFunctions'
-import { getMutableState, NO_PROXY, useMutableState } from '@etherealengine/hyperflux'
+import { defineComponent, getOptionalComponent, hasComponent } from '@etherealengine/ecs/src/ComponentFunctions'
+import { getMutableState, getState, NO_PROXY, useMutableState } from '@etherealengine/hyperflux'
+import { TransformComponent } from '@etherealengine/spatial'
 import { createTransitionState } from '@etherealengine/spatial/src/common/functions/createTransitionState'
 import { EngineState } from '@etherealengine/spatial/src/EngineState'
 import { InputComponent } from '@etherealengine/spatial/src/input/components/InputComponent'
+import { RigidBodyComponent } from '@etherealengine/spatial/src/physics/components/RigidBodyComponent'
 import { HighlightComponent } from '@etherealengine/spatial/src/renderer/components/HighlightComponent'
+import { VisibleComponent } from '@etherealengine/spatial/src/renderer/components/VisibleComponent'
 import { RendererComponent } from '@etherealengine/spatial/src/renderer/WebGLRendererSystem'
 import { BoundingBoxComponent } from '@etherealengine/spatial/src/transform/components/BoundingBoxComponents'
+import { ComputedTransformComponent } from '@etherealengine/spatial/src/transform/components/ComputedTransformComponent'
 import {
   DistanceFromCameraComponent,
   DistanceFromLocalClientComponent
 } from '@etherealengine/spatial/src/transform/components/DistanceComponents'
 import { EntityTreeComponent } from '@etherealengine/spatial/src/transform/components/EntityTree'
+import { XRUIComponent } from '@etherealengine/spatial/src/xrui/components/XRUIComponent'
+import { WebLayer3D } from '@etherealengine/xrui'
 import { useEffect } from 'react'
+import { MathUtils, Vector3 } from 'three'
 import matches from 'ts-matches'
+import { AvatarComponent } from '../../avatar/components/AvatarComponent'
 import { createUI } from '../functions/createUI'
-import { InteractableTransitions } from '../functions/interactableFunctions'
+import { inFrustum, InteractableState, InteractableTransitions } from '../functions/interactableFunctions'
 /**
  * Visibility override for XRUI, none is default behavior, on or off forces that state
  *
@@ -70,6 +79,98 @@ export enum XRUIActivationType {
   hover = 1
 }
 
+const xrDistVec3 = new Vector3()
+
+const updateXrDistVec3 = (selfAvatarEntity: Entity) => {
+  //TODO change from using rigidbody to use the transform position (+ height of avatar)
+  const selfAvatarRigidBodyComponent = getComponent(selfAvatarEntity, RigidBodyComponent)
+  const avatar = getComponent(selfAvatarEntity, AvatarComponent)
+  xrDistVec3.copy(selfAvatarRigidBodyComponent.position)
+  xrDistVec3.y += avatar.avatarHeight
+}
+
+export const updateInteractableUI = (entity: Entity) => {
+  const selfAvatarEntity = AvatarComponent.getSelfAvatarEntity()
+  const interactable = getComponent(entity, InteractableComponent)
+
+  if (!selfAvatarEntity || !interactable || interactable.uiEntity == UndefinedEntity) return
+
+  const xrui = getOptionalComponent(interactable.uiEntity, XRUIComponent)
+  const xruiTransform = getOptionalComponent(interactable.uiEntity, TransformComponent)
+  if (!xrui || !xruiTransform) return
+
+  const boundingBox = getOptionalComponent(entity, BoundingBoxComponent)
+
+  updateXrDistVec3(selfAvatarEntity)
+
+  const hasVisibleComponent = hasComponent(interactable.uiEntity, VisibleComponent)
+  if (hasVisibleComponent) {
+    TransformComponent.getWorldPosition(entity, xruiTransform.position)
+
+    //open to changing default height, 0.5 seems too small an offset (on default geo cube the xrui is half inside the cube if offset it just 0.5 from position)
+    xruiTransform.position.y += boundingBox ? 0.5 + boundingBox.box.max.y : 1
+
+    const cameraTransform = getComponent(Engine.instance.viewerEntity, TransformComponent)
+    xruiTransform.rotation.copy(cameraTransform.rotation)
+    xruiTransform.scale.set(1, 1, 1)
+  }
+
+  const distance = xrDistVec3.distanceToSquared(xruiTransform.position)
+
+  //slightly annoying to check this condition twice, but keeps distance calc on same frame
+  if (hasVisibleComponent) {
+    xruiTransform.scale.addScalar(MathUtils.clamp(distance * 0.01, 1, 5))
+  }
+
+  const transition = InteractableTransitions.get(entity)!
+  let activateUI = false
+
+  const inCameraFrustum = inFrustum(entity)
+  let hovering = false
+
+  if (inCameraFrustum) {
+    if (interactable.uiVisibilityOverride === XRUIVisibilityOverride.none) {
+      if (interactable.uiActivationType === XRUIActivationType.proximity) {
+        //proximity
+        let thresh = interactable.activationDistance
+        thresh *= thresh //squared for dist squared comparison
+        activateUI = distance < thresh
+      } else if (interactable.uiActivationType === XRUIActivationType.hover || interactable.clickInteract) {
+        //hover
+        const input = getOptionalComponent(entity, InputComponent)
+        if (input) {
+          hovering = input.inputSources.length > 0
+          activateUI = hovering
+        }
+      }
+    } else {
+      activateUI = interactable.uiVisibilityOverride !== XRUIVisibilityOverride.off //could be more explicit, needs to be if we add more enum options
+    }
+  }
+
+  //highlight if hovering OR if closest, otherwise turn off highlight
+  const mutableInteractable = getMutableComponent(entity, InteractableComponent)
+  mutableInteractable.highlighted.set(hovering || entity === getState(InteractableState).available[0])
+
+  if (transition.state === 'OUT' && activateUI) {
+    transition.setState('IN')
+    setComponent(interactable.uiEntity, VisibleComponent)
+  }
+  if (transition.state === 'IN' && !activateUI) {
+    transition.setState('OUT')
+  }
+  const deltaSeconds = getState(ECSState).deltaSeconds
+  transition.update(deltaSeconds, (opacity) => {
+    if (opacity === 0) {
+      removeComponent(interactable.uiEntity, VisibleComponent)
+    }
+    xrui.rootLayer.traverseLayersPreOrder((layer: WebLayer3D) => {
+      const mat = layer.contentMesh.material as THREE.MeshBasicMaterial
+      mat.opacity = opacity
+    })
+  })
+}
+
 /**
  * Adds an interactable UI to the entity if it has label text
  * @param entity
@@ -82,6 +183,10 @@ const addInteractableUI = (entity: Entity) => {
 
   interactable.uiEntity.set(createUI(entity, interactable.label.value, interactable.uiInteractable.value).entity)
   setComponent(interactable.uiEntity.value, EntityTreeComponent, { parentEntity: Engine.instance.originEntity })
+  setComponent(interactable.uiEntity.value, ComputedTransformComponent, {
+    referenceEntities: [entity, Engine.instance.viewerEntity],
+    computeFunction: () => updateInteractableUI(entity)
+  })
 
   const transition = createTransitionState(0.25)
   transition.setState('OUT')
