+++ packages/engine/src/interaction/systems/InteractableSystem.ts
@@ -24,19 +24,9 @@ Ethereal Engine. All Rights Reserved.
 */
 
 import { Not } from 'bitecs'
-import { MathUtils, Vector3 } from 'three'
 
-import { WebLayer3D } from '@etherealengine/xrui'
-
-import {
-  getMutableComponent,
-  hasComponent,
-  InputSystemGroup,
-  removeComponent,
-  UndefinedEntity,
-  UUIDComponent
-} from '@etherealengine/ecs'
-import { getComponent, getOptionalComponent, setComponent } from '@etherealengine/ecs/src/ComponentFunctions'
+import { hasComponent, InputSystemGroup, removeComponent, UndefinedEntity, UUIDComponent } from '@etherealengine/ecs'
+import { getComponent, getOptionalComponent } from '@etherealengine/ecs/src/ComponentFunctions'
 import { ECSState } from '@etherealengine/ecs/src/ECSState'
 import { Engine } from '@etherealengine/ecs/src/Engine'
 import { Entity } from '@etherealengine/ecs/src/Entity'
@@ -50,117 +40,19 @@ import { InputPointerComponent } from '@etherealengine/spatial/src/input/compone
 import { InputSourceComponent } from '@etherealengine/spatial/src/input/components/InputSourceComponent'
 import { XRStandardGamepadButton } from '@etherealengine/spatial/src/input/state/ButtonState'
 import { InputState } from '@etherealengine/spatial/src/input/state/InputState'
-import { RigidBodyComponent } from '@etherealengine/spatial/src/physics/components/RigidBodyComponent'
 import { HighlightComponent } from '@etherealengine/spatial/src/renderer/components/HighlightComponent'
-import { VisibleComponent } from '@etherealengine/spatial/src/renderer/components/VisibleComponent'
-import { BoundingBoxComponent } from '@etherealengine/spatial/src/transform/components/BoundingBoxComponents'
 import { DistanceFromCameraComponent } from '@etherealengine/spatial/src/transform/components/DistanceComponents'
-import { TransformComponent } from '@etherealengine/spatial/src/transform/components/TransformComponent'
 import { TransformSystem } from '@etherealengine/spatial/src/transform/systems/TransformSystem'
-import { XRUIComponent } from '@etherealengine/spatial/src/xrui/components/XRUIComponent'
 import { AvatarComponent } from '../../avatar/components/AvatarComponent'
-import { InteractableComponent, XRUIActivationType, XRUIVisibilityOverride } from '../components/InteractableComponent'
+import { InteractableComponent } from '../components/InteractableComponent'
 import {
   gatherAvailableInteractables,
-  inFrustum,
   InteractableState,
   InteractableTransitions
 } from '../functions/interactableFunctions'
 
 //TODO get rid of the query.exit and put it in component as unmount useEffect return
 //TODO get rid of this map eventually and store on the component instead
-
-const xrDistVec3 = new Vector3()
-
-const updateXrDistVec3 = (selfAvatarEntity: Entity) => {
-  //TODO change from using rigidbody to use the transform position (+ height of avatar)
-  const selfAvatarRigidBodyComponent = getComponent(selfAvatarEntity, RigidBodyComponent)
-  const avatar = getComponent(selfAvatarEntity, AvatarComponent)
-  xrDistVec3.copy(selfAvatarRigidBodyComponent.position)
-  xrDistVec3.y += avatar.avatarHeight
-}
-
-export const onInteractableUpdate = (entity: Entity) => {
-  const selfAvatarEntity = AvatarComponent.getSelfAvatarEntity()
-  const interactable = getComponent(entity, InteractableComponent)
-
-  if (!selfAvatarEntity || !interactable || interactable.uiEntity == UndefinedEntity) return
-
-  const xrui = getOptionalComponent(interactable.uiEntity, XRUIComponent)
-  const xruiTransform = getOptionalComponent(interactable.uiEntity, TransformComponent)
-  if (!xrui || !xruiTransform) return
-
-  const boundingBox = getOptionalComponent(entity, BoundingBoxComponent)
-
-  updateXrDistVec3(selfAvatarEntity)
-
-  const hasVisibleComponent = hasComponent(interactable.uiEntity, VisibleComponent)
-  if (hasVisibleComponent) {
-    TransformComponent.getWorldPosition(entity, xruiTransform.position)
-
-    //open to changing default height, 0.5 seems too small an offset (on default geo cube the xrui is half inside the cube if offset it just 0.5 from position)
-    xruiTransform.position.y += boundingBox ? 0.5 + boundingBox.box.max.y : 1
-
-    const cameraTransform = getComponent(Engine.instance.viewerEntity, TransformComponent)
-    xruiTransform.rotation.copy(cameraTransform.rotation)
-    xruiTransform.scale.set(1, 1, 1)
-  }
-
-  const distance = xrDistVec3.distanceToSquared(xruiTransform.position)
-
-  //slightly annoying to check this condition twice, but keeps distance calc on same frame
-  if (hasVisibleComponent) {
-    xruiTransform.scale.addScalar(MathUtils.clamp(distance * 0.01, 1, 5))
-  }
-
-  const transition = InteractableTransitions.get(entity)!
-  let activateUI = false
-
-  const inCameraFrustum = inFrustum(entity)
-  let hovering = false
-
-  if (inCameraFrustum) {
-    if (interactable.uiVisibilityOverride === XRUIVisibilityOverride.none) {
-      if (interactable.uiActivationType === XRUIActivationType.proximity) {
-        //proximity
-        let thresh = interactable.activationDistance
-        thresh *= thresh //squared for dist squared comparison
-        activateUI = distance < thresh
-      } else if (interactable.uiActivationType === XRUIActivationType.hover || interactable.clickInteract) {
-        //hover
-        const input = getOptionalComponent(entity, InputComponent)
-        if (input) {
-          hovering = input.inputSources.length > 0
-          activateUI = hovering
-        }
-      }
-    } else {
-      activateUI = interactable.uiVisibilityOverride !== XRUIVisibilityOverride.off //could be more explicit, needs to be if we add more enum options
-    }
-  }
-
-  //highlight if hovering OR if closest, otherwise turn off highlight
-  const mutableInteractable = getMutableComponent(entity, InteractableComponent)
-  mutableInteractable.highlighted.set(hovering || entity === getState(InteractableState).available[0])
-
-  if (transition.state === 'OUT' && activateUI) {
-    transition.setState('IN')
-    setComponent(interactable.uiEntity, VisibleComponent)
-  }
-  if (transition.state === 'IN' && !activateUI) {
-    transition.setState('OUT')
-  }
-  const deltaSeconds = getState(ECSState).deltaSeconds
-  transition.update(deltaSeconds, (opacity) => {
-    if (opacity === 0) {
-      removeComponent(interactable.uiEntity, VisibleComponent)
-    }
-    xrui.rootLayer.traverseLayersPreOrder((layer: WebLayer3D) => {
-      const mat = layer.contentMesh.material as THREE.MeshBasicMaterial
-      mat.opacity = opacity
-    })
-  })
-}
 
 const interactableQuery = defineQuery([InteractableComponent, Not(AvatarComponent), DistanceFromCameraComponent])
 const hoverInputInteractablesQuery = defineQuery([InteractableComponent, InputComponent])
@@ -181,10 +73,6 @@ const execute = () => {
 
   if (gatherAvailableInteractablesTimer === 0) {
     gatherAvailableInteractables(interactables)
-  }
-
-  for (const entity of interactables) {
-    onInteractableUpdate(entity)
   }
 }
