+++ packages/spatial/src/camera/systems/CameraFadeBlackEffectSystem.ts
@@ -40,10 +40,7 @@ import { addObjectToGroup } from '../../renderer/components/GroupComponent'
 import { setObjectLayers } from '../../renderer/components/ObjectLayerComponent'
 import { setVisibleComponent } from '../../renderer/components/VisibleComponent'
 import { ObjectLayers } from '../../renderer/constants/ObjectLayers'
-import {
-  ComputedTransformComponent,
-  setComputedTransformComponent
-} from '../../transform/components/ComputedTransformComponent'
+import { ComputedTransformComponent } from '../../transform/components/ComputedTransformComponent'
 import { TransformComponent } from '../../transform/components/TransformComponent'
 import { CameraActions } from '../CameraState'
 import { CameraSystem } from './CameraSystem'
@@ -66,10 +63,13 @@ const execute = () => {
   for (const action of fadeToBlackQueue()) {
     transition.setState(action.in ? 'IN' : 'OUT')
     if (action.in) {
-      setComputedTransformComponent(entity, Engine.instance.cameraEntity, () => {
-        getComponent(entity, TransformComponent).position.copy(
-          getComponent(Engine.instance.cameraEntity, TransformComponent).position
-        )
+      setComponent(entity, ComputedTransformComponent, {
+        referenceEntities: [Engine.instance.cameraEntity],
+        computeFunction: () => {
+          getComponent(entity, TransformComponent).position.copy(
+            getComponent(Engine.instance.cameraEntity, TransformComponent).position
+          )
+        }
       })
     } else removeComponent(entity, ComputedTransformComponent)
