+++ packages/spatial/src/transform/systems/TransformSystem.ts
@@ -68,7 +68,7 @@ declare module 'three/src/core/Object3D' {
 
 export const computeTransformMatrix = (entity: Entity) => {
   const transform = getComponent(entity, TransformComponent)
-  updateTransformFromComputedTransform(entity)
+  getOptionalComponent(entity, ComputedTransformComponent)?.computeFunction()
   composeMatrix(entity)
   const entityTree = getOptionalComponent(entity, EntityTreeComponent)
   const parentEntity = entityTree?.parentEntity
@@ -78,12 +78,6 @@ export const computeTransformMatrix = (entity: Entity) => {
   } else {
     transform.matrixWorld.copy(transform.matrix)
   }
-}
-
-const updateTransformFromComputedTransform = (entity: Entity) => {
-  const computedTransform = getOptionalComponent(entity, ComputedTransformComponent)
-  if (!computedTransform) return
-  computedTransform.computeFunction(entity, computedTransform.referenceEntity)
 }
 
 export const updateGroupChildren = (entity: Entity) => {
@@ -108,33 +102,17 @@ export const getDistanceSquaredFromTarget = (entity: Entity, targetPosition: Vec
 const _frustum = new Frustum()
 const _projScreenMatrix = new Matrix4()
 
-const originChildEntities = new Set<Entity>()
-
-/** get list of entities that are children of the world origin */
-const updateOriginChildEntities = (entity: Entity) => {
-  const referenceEntity = getOptionalComponent(entity, ComputedTransformComponent)?.referenceEntity
-  const parentEntity = getOptionalComponent(entity, EntityTreeComponent)?.parentEntity
-
-  if (
-    referenceEntity &&
-    (originChildEntities.has(referenceEntity) || referenceEntity === Engine.instance.localFloorEntity)
-  )
-    originChildEntities.add(referenceEntity)
-  if (parentEntity && (originChildEntities.has(parentEntity) || parentEntity === Engine.instance.localFloorEntity))
-    originChildEntities.add(parentEntity)
-}
-
 const transformDepths = new Map<Entity, number>()
 
 const updateTransformDepth = (entity: Entity) => {
   if (transformDepths.has(entity)) return transformDepths.get(entity)
 
-  const referenceEntity = getOptionalComponent(entity, ComputedTransformComponent)?.referenceEntity
+  const referenceEntities = getOptionalComponent(entity, ComputedTransformComponent)?.referenceEntities
   const parentEntity = getOptionalComponent(entity, EntityTreeComponent)?.parentEntity
 
-  const referenceEntityDepth = referenceEntity ? updateTransformDepth(referenceEntity) : 0
+  const referenceEntityDepths = referenceEntities ? referenceEntities.map(updateTransformDepth) : []
   const parentEntityDepth = parentEntity ? updateTransformDepth(parentEntity) : 0
-  const depth = Math.max(referenceEntityDepth, parentEntityDepth) + 1
+  const depth = Math.max(...referenceEntityDepths, parentEntityDepth) + 1
   transformDepths.set(entity, depth)
 
   return depth
@@ -177,21 +155,17 @@ const execute = () => {
   if (needsSorting) {
     transformDepths.clear()
     for (const entity of sortedTransformEntities) updateTransformDepth(entity)
-    for (const entity of sortedTransformEntities) updateOriginChildEntities(entity)
     insertionSort(sortedTransformEntities, compareReferenceDepth) // Insertion sort is speedy O(n) for mostly sorted arrays
     TransformComponent.transformsNeedSorting = false
   }
 
   // entities with dirty parent or reference entities, or computed transforms, should also be dirty
   for (const entity of sortedTransformEntities) {
-    const makeDirty =
+    TransformComponent.dirtyTransforms[entity] =
       TransformComponent.dirtyTransforms[entity] ||
+      hasComponent(entity, ComputedTransformComponent) ||
       TransformComponent.dirtyTransforms[getOptionalComponent(entity, EntityTreeComponent)?.parentEntity ?? -1] ||
-      TransformComponent.dirtyTransforms[
-        getOptionalComponent(entity, ComputedTransformComponent)?.referenceEntity ?? -1
-      ] ||
-      hasComponent(entity, ComputedTransformComponent)
-    TransformComponent.dirtyTransforms[entity] = makeDirty
+      false
   }
 
   const dirtySortedTransformEntities = sortedTransformEntities.filter(isDirty)
@@ -249,8 +223,6 @@ const reactor = () => {
 
     return () => {
       networkState.networkSchema[TransformSerialization.ID].set(none)
-
-      originChildEntities.clear()
       sortedTransformEntities.length = 0
       transformDepths.clear()
     }
